Move axes logic to new Axes class
